requirements.txt
pip
setuptools
wheel
numpy
pandas
scipy
torch
torchvision
torchmetrics
ml_collections
segmentation_models_pytorch
einops
matplotlib
librosa
soundfile
demucs
audiomentations
torch_audiomentations
httpx
gradio
google-api-python-client
pytube
tqdm
psutil
omegaconf
beartype
asteroid
torchseg
yt_dlp
rotary_embedding_torch
noisereduce
loralib
hyper_connections==0.1.11
pyngrok
matchering==2.0.6
pyloudnorm
gdown
validators
requests
sympy
typing_extensions
packaging
opencv-python-headless
pypresence
mpmath
pillow
